Refactor Tab component exports/imports
