+++ src/components/tabs/index.js
@@ -1,0 +1,2 @@
+export { default as Tabs } from "components/tabs/Tabs";
+export { default as Tab } from "components/tabs/Tab";
